% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_color.R
\name{closest_color}
\alias{closest_color}
\title{Find most similar built-in R color}
\usage{
closest_color(hex, n = 1, print = T)
}
\arguments{
\item{hex}{hex color code}

\item{n}{top n closest colors}

\item{print}{do print colors side-by-side?}
}
\value{
color name
}
\description{
Find most similar built-in R color
}
\examples{
colrr::closest_color(hex =  rainbow(10))
}
