% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_scale_continuous.R
\name{get_color_scale_continuous}
\alias{get_color_scale_continuous}
\title{Make a continuous fill/color scale for ggplot}
\usage{
get_color_scale_continuous(
  values,
  zscored = NULL,
  colorsteps = "..auto..",
  legendbreaks = "..auto..",
  legendlabels = "..auto..",
  colors = col_pal("RdBu", direction = -1),
  colorsteps_nice = T,
  type = c("fill", "color"),
  col_na = "grey50",
  qmin = 0,
  qmax = 1,
  scale.max = NULL,
  scale.min = NULL
)
}
\arguments{
\item{values}{vector of numeric values}

\item{zscored}{are values z-scored? if NULL, evaluated within function}

\item{colorsteps}{NULL for colorbar, a number for number of steps or ..auto..
for decision based on zscored etc; when colorsteps_nice = T it influences number
of steps}

\item{legendbreaks}{vector of breaks}

\item{legendlabels}{vector of labels}

\item{colors}{to ggplot2::scale_fill/color_gradientn or
ggplot2::scale_fill/color_stepsn}

\item{colorsteps_nice}{to nice.breaks in ggplot2::scale_fill/color_stepsn}

\item{type}{fill or color type to return}

\item{col_na}{color for NA values}

\item{qmin}{adjust legend to a lower quantile cutoff}

\item{qmax}{adjust legend to an upper quantile cutoff}

\item{scale.max}{formatted (e.g. rounded) max of values; if NULL, calculated
within function}

\item{scale.min}{see scale.max}
}
\value{
ggplot scale object
}
\description{
Make nice default color steps.
}
