% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_contrast_subset.R
\name{get_max_contrast_subset}
\alias{get_max_contrast_subset}
\title{Get a subset of colors with max contrasts (simple)}
\usage{
get_max_contrast_subset(
  cols,
  contrast_matrix = NULL,
  k = 5,
  seed = 42,
  include = NULL
)
}
\arguments{
\item{cols}{color vector}

\item{contrast_matrix}{optional pre-computed contrast matrix with
colrr::contrast_ratio(cols, cols)}

\item{k}{number of colors to return}

\item{seed}{seed for random start}

\item{include}{indices in cols to keep}
}
\value{
subset color vector from cols
}
\description{
Get a subset of colors with max contrasts (simple)
}
\examples{
get_max_contrast_subset(colrr::col_pal("material"))
}
