% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_contrast_subset2.R
\name{get_max_contrast_subset2}
\alias{get_max_contrast_subset2}
\title{Get a subset of colors with max contrasts}
\usage{
get_max_contrast_subset2(
  cols,
  contrast_matrix = NULL,
  k = 5,
  include = NULL,
  exclude = NULL,
  topn = 1
)
}
\arguments{
\item{cols}{color vector}

\item{contrast_matrix}{optional pre-computed contrast matrix with
colrr::contrast_ratio(cols, cols)}

\item{k}{number of colors to return}

\item{include}{indices in cols to keep}

\item{exclude}{indices in cols to remove}

\item{topn}{number of top results to return}
}
\value{
subset color vector from cols
}
\description{
Get a subset of colors with max contrasts
}
\examples{
get_max_contrast_subset2(colrr::col_pal("material"))
}
