% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_material.R
\name{guides_default}
\alias{guides_default}
\title{Add default legend guide elements to ggplot}
\usage{
guides_default(colorbar = F, ...)
}
\arguments{
\item{colorbar}{plot color legend as bar}

\item{...}{arguments to guide_colorbar or guide_legend}
}
\value{
ggplot2 guide
}
\description{
Make default barheight = 10 and width = 1, legend title as
ggtext::element_markdown, point legends of size = 4. Valid for fill and color.
}
\examples{
df <- data.frame(x = rnorm(100), y = rnorm(100), class = sample(letters[1:10], 100, T))
gplot <- ggplot2::ggplot(df, ggplot2::aes(x,y)) +
  ggplot2::geom_point(ggplot2::aes(color = class)) +
  ggplot2::scale_color_manual(values = colrr::col_pal("material"))
# larger dots in legend per default
gplot + colrr::theme_material() + guides_default()
df2 <- data.frame(class = letters[1:3], y = c(20,30,40))
gplot <- ggplot2::ggplot(df2, ggplot2::aes(class,y)) +
  ggplot2::geom_col(ggplot2::aes(fill = y))
# larger bar by default
gplot  + colrr::theme_material() + guides_default(colorbar = T)
}
