% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_grey.R
\name{hex_to_grey}
\alias{hex_to_grey}
\title{Hey color to grey scale}
\usage{
hex_to_grey(hex, out = c("grey", "bw"), bw_thresh = 0.5, grey_algo = c(3, 2))
}
\arguments{
\item{hex}{hex color vector}

\item{out}{output}

\item{bw_thresh}{threshold for black/white}

\item{grey_algo}{algorithm for greyscale}
}
\description{
Hey color to grey scale
}
