% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_col_pal.R
\name{make_col_pal}
\alias{make_col_pal}
\title{Algorithm to adjust color palette to factor levels which are to be colored}
\usage{
make_col_pal(
  col_vec,
  fct_lvls = NULL,
  col_pal_args = list(),
  adjust_to_fct_lvls = F,
  na_color = "grey50",
  missing_fct_to_na = T
)
}
\arguments{
\item{col_vec}{named or unnamed, complete or incomplete vector of colors;
names should correspond to fct_lvls, can be an R color name, a hex color
starting with '#' or a vector of these}

\item{fct_lvls}{factor levels to color, can be a character; NULL for
continuous scale w/o factor levels to color}

\item{col_pal_args}{arguments to colrr::col_pal but not n}

\item{adjust_to_fct_lvls}{adjust missing or wrong names of col_pal?}

\item{na_color}{color for NA}

\item{missing_fct_to_na}{add NA for missing colors?}
}
\value{
named color vector
}
\description{
Heuristic to handle shorter, longer, named, unnamed col_vec in comparison to
fct_lvls.
}
\examples{
exdata <- readRDS(system.file("extdata", "make_col_pal_example.rds", package = "colrr"))
ex_col <- exdata[[1]]
fct_lvls <- exdata[[2]]
# exact col_vec
make_col_pal(col_vec = ex_col[unique(fct_lvls)], fct_lvls = unique(fct_lvls))
# longer col_vec
make_col_pal(col_vec = ex_col, fct_lvls = unique(fct_lvls))
# dup fct_lvls
make_col_pal(col_vec = ex_col, fct_lvls = fct_lvls)
# longer col_vec unnamed
make_col_pal(col_vec = unname(ex_col), fct_lvls = unique(fct_lvls))
# shorter col_vec
make_col_pal(col_vec = ex_col[1:5], fct_lvls = unique(fct_lvls))
make_col_pal(col_vec = unname(ex_col[1:5]), fct_lvls = unique(fct_lvls))
# with adjust - hue pal
make_col_pal(col_vec = ex_col[1:5], fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = T)
# other order
make_col_pal(col_vec = ex_col[1:5], fct_lvls = sort(unique(fct_lvls)), adjust_to_fct_lvls = T)
# only "" names (= no matching name)
make_col_pal(col_vec = stats::setNames(ex_col, rep("", length(ex_col))), fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = F)
make_col_pal(col_vec = stats::setNames(ex_col, rep("", length(ex_col))), fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = T)
# some "" names
ex_col2 <- ex_col
names(ex_col2)[c(3,5,7,9)] <- ""
make_col_pal(col_vec = ex_col2, fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = T)
make_col_pal(col_vec = ex_col2, fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = F)
# some wrong names
names(ex_col2)[c(3,5,7,9)] <- c(letters[1:4])
make_col_pal(col_vec = ex_col2, fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = T)
make_col_pal(col_vec = ex_col2, fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = F)
# single r color
make_col_pal(col_vec = "grey90", fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = F)
make_col_pal(col_vec = "grey90", fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = T)
# single hex color
make_col_pal(col_vec = "#FFB90FFF", fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = F)
make_col_pal(col_vec = "#FFB90FFF", fct_lvls = unique(fct_lvls), adjust_to_fct_lvls = T)

# then use:
# ggplot2::scale_color_manual(values = col.pal,
#                             na.value = col_na)
}
