% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_colors.R
\name{order_colors}
\alias{order_colors}
\title{Title}
\usage{
order_colors(x, method = c("TSP", "hclust", "path"))
}
\description{
Title
}
\examples{
rcolors <- unique(grDevices::colors())
rcolorsrgb <- grDevices::col2rgb(rcolors, alpha = T)
rcolorshex <- grDevices::rgb(rcolorsrgb[1, ], rcolorsrgb[2, ], rcolorsrgb[3, ], maxColorValue = 255, alpha = T)
out <- order_colors(rcolorsrgb)
scales::show_col(rcolors[out], labels = F)
}
