% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_material.R
\name{theme_material}
\alias{theme_material}
\title{ggplot2 theme with style of Rstudios color palette 'Material'}
\usage{
theme_material(
  base_size = 14,
  text_color = "white",
  bg_color = "#273238",
  bg_color2 = "#4F5C69",
  text_fun = ggtext::element_markdown,
  style = c("prism", "prismy", "prismx", "base"),
  white = F,
  legend_tight = F,
  ...
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{text_color}{color of various text elements}

\item{bg_color}{background color, not intended to be changed dramatically}

\item{bg_color2}{background color 2, e.g. for facet strip}

\item{text_fun}{function for text elements, ggplot2::element_text or
ggtext::element_markdown}

\item{style}{appearance of axes and grid}

\item{white}{use a light color scale with white bg}

\item{legend_tight}{legend more compact and closer to plot panel}
}
\value{
ggplot2 theme
}
\description{
Geoms need to be turned to white (or bright) coloring.
}
\examples{
gplot <- ggplot2::ggplot(data.frame(x = rnorm(100), y = rnorm(100), class = sample(letters[1:10], 100, T)), ggplot2::aes(x,y)) +
  ggplot2::geom_point(ggplot2::aes(color = class)) +
  ggplot2::scale_color_manual(values = colrr::col_pal("material"))
gplot + colrr::theme_material()
gplot + colrr::theme_material(white = T)
gplot + colrr::theme_material(style = "prismy")
gplot + colrr::theme_material(text_color = "hotpink", style = "base")
}
