% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_to_hex.R
\name{wl_to_hex}
\alias{wl_to_hex}
\title{Wavelength to hex color value}
\usage{
wl_to_hex(wavelength, gamma = 0.8)
}
\arguments{
\item{wavelength}{wl in 380 nm to 780 nm}

\item{gamma}{delta}
}
\value{
hex character
}
\description{
From this old repo https://github.com/ucl-cssb/wavelength2colour to chatgpt
to here. Can it be improved?
}
\examples{
scales::show_col(wl_to_hex(seq(380,780,2)))
}
