% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scale_fun.R
\name{get_scale_fun}
\alias{get_scale_fun}
\alias{get_scale_color_fun}
\alias{get_scale_fill_fun}
\title{Make a continuous fill/color scale for ggplot}
\usage{
get_scale_fun(
  values,
  zscored = NULL,
  steps = "..auto..",
  legendbreaks = "..auto..",
  legendlabels = "..auto..",
  palette = col_pal("RdBu", direction = -1),
  steps_nice = T,
  type = c("fill", "color"),
  col_na = "grey50",
  qmin = 0,
  qmax = 1,
  scale_max = NULL,
  scale_min = NULL,
  trans_log = F,
  center_zero = F,
  ...
)

get_scale_color_fun(
  values,
  zscored = NULL,
  steps = "..auto..",
  legendbreaks = "..auto..",
  legendlabels = "..auto..",
  palette = col_pal("RdBu", direction = -1),
  steps_nice = T,
  col_na = "grey50",
  qmin = 0,
  qmax = 1,
  scale_max = NULL,
  scale_min = NULL,
  trans_log = F,
  center_zero = F,
  ...
)

get_scale_fill_fun(
  values,
  zscored = NULL,
  steps = "..auto..",
  legendbreaks = "..auto..",
  legendlabels = "..auto..",
  palette = col_pal("RdBu", direction = -1),
  steps_nice = T,
  col_na = "grey50",
  qmin = 0,
  qmax = 1,
  scale_max = NULL,
  scale_min = NULL,
  trans_log = F,
  center_zero = F,
  ...
)
}
\arguments{
\item{values}{vector of numeric values}

\item{zscored}{are values z-scored? if NULL, evaluated within function}

\item{steps}{NULL for colorbar, a number for number of steps or a vector or
..auto.. for decision based on zscored etc; when steps_nice = T it
influences number of steps}

\item{legendbreaks}{vector of breaks}

\item{legendlabels}{vector of labels}

\item{palette}{to color arg of ggplot2::scale_fill/color_gradientn or
ggplot2::scale_fill/color_stepsn}

\item{steps_nice}{to nice.breaks in ggplot2::scale_fill/color_stepsn}

\item{type}{fill or color type to return}

\item{col_na}{color for NA values}

\item{qmin}{adjust legend by percentile cutoff at lower end}

\item{qmax}{adjust legend by percentile cutoff at upper end}

\item{scale_max}{formatted (e.g. rounded) max of values; if NULL, calculated
within function}

\item{scale_min}{see max}

\item{trans_log}{log transform colorscale?}
}
\value{
ggplot scale object
}
\description{
Make nice default color steps.
}
\examples{
library(ggplot2)
library(colrr)

df <- data.frame(x = rnorm(1000),
                 y = rnorm(1000),
                 z = rnorm(1000))

p <- ggplot(df, aes(x,y,color = z)) +
  geom_point()
p

# color steps by default
p + colrr::get_scale_color_fun(df$z)

# back to normal
p + colrr::get_scale_color_fun(df$z,
                               steps = NULL)

# set number of steps
p + colrr::get_scale_color_fun(df$z,
                               steps = 12)

# set steps exactly
p + colrr::get_scale_color_fun(df$z,
                               steps = c(-2.5,-2,0,1))

# other palette
p + colrr::get_scale_color_fun(df$z,
                               palette = colorRamps::blue2red(n = 50))



# color vecor not z-scored
# colpal_info <- col_pal()
df$z <- runif(1000, max = 9)
p <- ggplot(df, aes(x,y,color = z)) +
  geom_point()

p + colrr::get_scale_color_fun(df$z,
                               palette = colrr::col_pal("spectral", direction = -1))

# spectral color palette from rcolorbrewer
p + colrr::get_scale_color_fun(df$z,
                               palette = rev(RColorBrewer::brewer.pal(11, "Spectral")))

# rainbow palette
p + colrr::get_scale_color_fun(df$z,
                               palette = grDevices::rainbow(n = 50),
                               steps = 10)
# jet palette
p + colrr::get_scale_color_fun(df$z,
                               palette = colrr::col_pal("jet"),
                               steps = NULL)

p + colrr::get_scale_color_fun(df$z,
                               palette = colrr::col_pal("jetColors"),
                               steps = NULL)

p + colrr::get_scale_color_fun(df$z,
                               palette = colrr::col_pal("spectral", direction = -1),
                               steps = NULL)

p + colrr::get_scale_color_fun(df$z,
                               palette = colrr::col_pal("spectral", direction = -1),
                               steps = 8)

# quantile limits: when cutting the color value range to any percentile, add this info to the color legend
df$z <- scales::squish(df$z, range = c(quantile(df$z, 0.2), quantile(df$z, 0.8)))
max(df$z)
min(df$z)
p <- ggplot(df, aes(x,y,color = z)) +
  geom_point()

p + colrr::get_scale_color_fun(df$z,
                               palette = colrr::col_pal("spectral", direction = -1),
                               qmin = 0.2,
                               qmax = 0.8)

# upper and lower limit missing yet, here
p + colrr::get_scale_color_fun(df$z,
                               palette = colrr::col_pal("spectral", direction = -1),
                               qmin = 0.2,
                               qmax = 0.8,
                               steps = NULL)

p + colrr::get_scale_color_fun(df$z,
                               palette = colrr::col_pal("spectral", direction = -1),
                               qmin = 0.2,
                               qmax = 0.8,
                               steps = 10)
}
